Log4j2: moving to XML representation.
Log4j2 does not work that well with modules.
It's impossible to access the provided JSON Layout.
The YAML configuration was a bit overkill for reading only 1 file.
Log4j3 is not stable. The latest version (a beta) has known bugs,
and I did not succeed to work them around. XML configurations works
fine, we'll see for JSON layout if really needed one day.